# docs/swagger.yaml
basePath: /
definitions:
...
  httptransport.DeliveryResponseBody:
    properties:
      application_id:
        type: string
      attempts:
        type: integer
      channel:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      failure_reason:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      locale:
        type: string
      next_attempt_at:
        type: string
      provider:
        type: string
      provider_message_id:
        type: string
      recipient:
        type: string
      status:
        type: string
      template_code:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      variables: {}
      version:
        type: integer
    type: object
...
  httptransport.GetProviderRequest:
    properties:
      application_id:
        type: string
      code:
        type: string
      tenant_id:
        type: string
    required:
    - application_id
    - code
    - tenant_id
    type: object
  httptransport.GetTemplateRequest:
    properties:
      application_id:
        type: string
      channel:
        type: string
      code:
        type: string
      locale:
        type: string
      tenant_id:
        type: string
    required:
    - application_id
    - channel
    - code
    - locale
    - tenant_id
    type: object
...
  httptransport.ListProvidersRequest:
    properties:
      application_id:
        type: string
      channel:
        type: string
      keyword:
        type: string
      page:
        type: integer
      page_size:
        type: integer
      status:
        type: string
      tenant_id:
        type: string
    required:
    - application_id
    - tenant_id
    type: object
  httptransport.ListTemplatesRequest:
    properties:
      application_id:
        type: string
      channel:
        type: string
      keyword:
        type: string
      page:
        type: integer
      page_size:
        type: integer
      status:
        type: string
      tenant_id:
        type: string
    required:
    - application_id
    - tenant_id
    type: object
  httptransport.MeResponseBody:
    properties:
      subject:
        type: string
    type: object
  httptransport.ProviderPageResponseBody:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      providers:
        items:
          $ref: '#/definitions/httptransport.ProviderResponseBody'
        type: array
      total:
        type: integer
    type: object
  httptransport.ProviderReceiptRequest:
    properties:
      application_id:
        type: string
      failure_reason:
        type: string
      provider:
        type: string
      provider_message_id:
        type: string
      status:
        type: string
      tenant_id:
        type: string
    required:
    - application_id
    - provider
    - provider_message_id
    - status
    - tenant_id
    type: object
  httptransport.ProviderResponseBody:
    properties:
      application_id:
        type: string
      channel:
        type: string
      code:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      path:
        type: string
      priority:
        type: integer
      status:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      upstream:
        type: string
      version:
        type: integer
    type: object
  httptransport.PutProviderRequest:
    properties:
      application_id:
        type: string
      channel:
        type: string
      code:
        type: string
      expected_version:
        type: integer
      path:
        type: string
      priority:
        type: integer
      status:
        type: string
      tenant_id:
        type: string
      upstream:
        type: string
    required:
    - application_id
    - channel
    - code
    - path
    - tenant_id
    - upstream
    type: object
  httptransport.PutTemplateRequest:
    properties:
      application_id:
        type: string
      channel:
        type: string
      code:
        type: string
      content:
        type: string
      expected_version:
        type: integer
      locale:
        type: string
      status:
        type: string
      subject:
        type: string
      tenant_id:
        type: string
    required:
    - application_id
    - channel
    - code
    - content
    - locale
    - tenant_id
    type: object
  httptransport.Response:
    properties:
      body: {}
      code:
        type: integer
      message:
        type: string
      request_id:
        type: string
    type: object
  httptransport.SendNotificationRequest:
    properties:
      application_id:
        type: string
      channel:
        type: string
      idempotency_key:
        type: string
      locale:
        type: string
      recipient:
        type: string
      template_code:
        type: string
      tenant_id:
        type: string
      variables:
        additionalProperties:
          type: string
        type: object
    required:
    - application_id
    - channel
    - idempotency_key
    - locale
    - recipient
    - template_code
    - tenant_id
    type: object
  httptransport.TemplatePageResponseBody:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      templates:
        items:
          $ref: '#/definitions/httptransport.TemplateResponseBody'
        type: array
      total:
        type: integer
    type: object
  httptransport.TemplateResponseBody:
    properties:
      application_id:
        type: string
      channel:
        type: string
      code:
        type: string
      content:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      locale:
        type: string
      status:
        type: string
      subject:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      version:
        type: integer
    type: object
info:
  contact: {}
  description: 'Production-oriented Go Web API scaffold. Application error codes:
    0 success; 10000-19999 common/input; 20000-29999 authentication/authorization;
    30000-39999 business; 50000-59999 infrastructure.'
  title: Go API Template
  version: "1.0"
paths:
  /api/v1/me:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.MeResponseBody'
              type: object
        "401":
          description: 'Code 20001: unauthorized'
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Return the authenticated subject
      tags:
      - authentication
...
  /api/v1/notifications/providers/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: Provider identity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.GetProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.ProviderResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Get a notification provider route
      tags:
      - notifications
  /api/v1/notifications/providers/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Provider filters and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.ListProvidersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.ProviderPageResponseBody'
              type: object
      security:
      - Bearer: []
      summary: List notification provider routes
      tags:
      - notifications
  /api/v1/notifications/providers/put:
    post:
      consumes:
      - application/json
      parameters:
      - description: Provider route
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.PutProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.ProviderResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Create or update a notification provider route
      tags:
      - notifications
  /api/v1/notifications/providers/receipt:
    post:
      parameters:
      - description: Provider receipt
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.ProviderReceiptRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.DeliveryResponseBody'
              type: object
      security:
      - PSK: []
      summary: Record an authenticated provider delivery receipt
      tags:
      - notification
  /api/v1/notifications/send:
    post:
      parameters:
      - description: Notification
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.SendNotificationRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.DeliveryResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Queue a notification for asynchronous delivery
      tags:
      - notification
  /api/v1/notifications/templates/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: Template identity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.GetTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.TemplateResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Get a notification template
      tags:
      - notification
  /api/v1/notifications/templates/list:
    post:
      parameters:
      - description: Template filter
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.ListTemplatesRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.TemplatePageResponseBody'
              type: object
      security:
      - Bearer: []
      summary: List notification templates
      tags:
      - notification
  /api/v1/notifications/templates/put:
    post:
      parameters:
      - description: Template
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.PutTemplateRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.TemplateResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Create or update notification template
      tags:
      - notification
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Enter "Bearer {JWT}".
    in: header
    name: Authorization
    type: apiKey
  PSK:
    description: Enter "PSK {shared-key}" for routes configured with PSK authentication.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"